Fix Metroid Room 1 by adding an extra node
Removed the dependence on the gray door strats to make the logic more generic so randomizers can more easily change the gray doors.

The logic in this room cannot simply be moved to the strats because there are multiple ways to kill or avoid the Metroids as well as multiple strats to traverse the room. 

In order to remedy both solutions, Node 3 and Obstacle A were added. 
- The player will traverse the room from 1->3->2 or 2->3->1. 
- Getting to node 3 represents killing or avoiding the Metroids. 
- Leaving Node 3 will represent crossing the room. 
- Killing the Metroids or having the metroids already dead (f_KilledMetroidRoom1) will clear obstacle A, which will unlock the gray door and yield the same flag.