Split the Node into Two
There is now a second node to make the room logic more accurate. 
The previous method was inaccurate when entering the room, killing the metroids, and leaving the room from the original door without the tech required to cross the acid for free. The reason is because this strat would require the player to cross the acid twice, where the logic would have only had them cross it once.