Add canWalljump to Pink Brinstar Hopper Room
This required a major overhaul of the room. In order to prevent a large number of combinations of ways to kill the hoppers, open the blue gate, and get over the central barrier, obstacle removal strats were added on single nodes. For example: killing the hoppers is a strat from node 1->1, but it clears an obstacle. This also allowed the removal of node 3.