added requirements for slopespark
As per discussions with Kyle & Maddo.. hopefully this is correct (or close to it!)... I added "canHorizontalDamageboost" as a
 requirement to the tech, which should implicitly require "canUseEnemies"

For the 2 slopes parks in spore spawn farming I added "CanShinesparkMovement" as it is only a simple left to right walk, no jumping etc unlike some other rooms so I think it represents this strat best? (other rooms like LN Firefleas slopespark off the boulder would use canShinechargeMovementComplex as it requires jumping / moonwalking etc..

You can only use the first slope as the shineframes run out for the 2nd one (I was mixed up with testing Aqua's idea of using a flashsuit to regain a flashsuit and leave with initial spark".

I reduced the availabletilecount by 1 as if you stop on the tile next to the pipe you take an hit from the zeb.