add canRmodeSpark to logic.
The second way of gaining a flashsuit without morphball, using an r-mode standup.

Video for added strat has Id: 8453.

As far as i know it doesn't matter what energy is in the reserve tank (between 1-4) the spark is always frame perfect (3rd frame after enemy contact, same frame as when the garbled graphics appear).

It's overall "easier" than slope spark but if you mess it up its a longer time to reset.. and in the case of slow moving enemies like atomics, if you spark out of the flashsuit too late it will kill you when iframes expire.

However, I still think it should be added to logic as it is part of the flash suit tech and can be used in rooms with a runway without slopes and without morph.