add spikesuits to Draygons room
I gave 4 options, with / without draygon defeated and with / without breaking a turret.

draygon alive / no ammo is pretty terrible, you have to dodge the projectiles and precise walljump / morph into the already broken turret.

https://videos.maprando.com/video/9160 [defeated / break turret]
https://videos.maprando.com/video/9159 [defeated / no break turret]
https://videos.maprando.com/video/9158 [alive / break turret]
https://videos.maprando.com/video/9157 [alive / no break turret]

Not 100% if all the tech / requirements are right but tried to put stuff onto the draygon alive / no ammo to ensure it gets placed at a higher difficulty as it is really tight and took many attempts to get it so I hope the randomizer wouldnt just put it there as an easy trick.